+++ packages/react/src/components/Container/Container.tsx
@@ -1,10 +1,11 @@
+import clsx from 'clsx';
 import styles from './Container.module.scss'
 
 import type { ContainerProps } from './Container.types';
 
-export default function Container({ children }: ContainerProps) {
+export default function Container({ children, className }: ContainerProps) {
   return (
-    <main className={`${styles['container']} mx-auto`}>
+    <main className={clsx(styles['container'], 'mx-auto', className)}>
         {children}
     </main>
   )