+++ packages/react/src/types/scss.d.ts
@@ -9,6 +9,11 @@ declare module '*.module.scss' {
   export default classes;
 }
 
+declare module '*.css' {
+  const classes: { [key: string]: string };
+  export default classes;
+}
+
 declare module '*.sass' {
   const content: { [className: string]: string };
   export default content;