+++ apps/storybook-react/CHANGELOG.md
@@ -1,5 +1,37 @@
 # storybook-react
 
+## 1.0.24
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @giro-ds/react@9.0.0
+
+## 1.0.23
+
+### Patch Changes
+
+- Updated dependencies [8adb857]
+- Updated dependencies [8adb857]
+- Updated dependencies [8adb857]
+  - @giro-ds/react@8.0.0
+
+## 1.0.22
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @giro-ds/react@7.0.0
+
 ## 1.0.21
 
 ### Patch Changes